@@ -16,6 +16,7 @@
  */
 const { Universal, MemoryAccount, Node, Crypto } = require( '@aeternity/aepp-sdk' )
 const contractUtils = require( '../utils/contract-utils' )
+const fs = require( 'fs' )
 
 const NETWORKS = require( '../config/network.json' )
 const DEFAULT_NETWORK_NAME = 'local'
@@ -34,13 +35,15 @@ const deploy = async ( secretKey, network, compiler ) => {
     }
     const NETWORK_NAME = network ? network : DEFAULT_NETWORK_NAME
 
-    const client = await Universal( {
+    const client = await Universal.compose( {
+        deepProps: { Ae: { defaults: { interval: 10 } } }
+    } )( {
         nodes: [
             {
-                name     : NETWORK_NAME, instance : await Node( { 
+                name     : NETWORK_NAME, instance : await Node( {
                     url           : NETWORKS[NETWORK_NAME].nodeUrl,
                     ignoreVersion : true
-                } ) 
+                } )
             },
         ],
         compilerUrl : compiler ? compiler : NETWORKS[NETWORK_NAME].compilerUrl,
@@ -49,7 +52,7 @@ const deploy = async ( secretKey, network, compiler ) => {
     } )
     // a filesystem object must be passed to the compiler if the contract uses custom includes
 
-    const deployContract = async ( source, params ) => {
+    const deployContract = async ( source, params, interfaceName ) => {
         try {
             console.log( '----------------------------------------------------------------------------------------------------' )
             console.log( `%cdeploying '${source}...'`, `color:green` )
@@ -57,28 +60,42 @@ const deploy = async ( secretKey, network, compiler ) => {
             const filesystem       = contractUtils.getFilesystem( source )
             const contract_content = contractUtils.getContractContent( source )
 
-            const contract          = await client.getContractInstance( contract_content, { filesystem } )
+            const contract          = await client.getContractInstance( { source: contract_content, filesystem } )
             const deployment_result = await contract.deploy( params )
             console.log( deployment_result )
             console.log( '-------------------------------------  END  ---------------------------------------------------------' )
+
+            if ( interfaceName ) {
+                const parent = "./contracts/interfaces/for-export"
+                if ( !fs.existsSync( parent ) ) {
+                    fs.mkdirSync( parent )
+                }
+                fs.writeFileSync( parent + '/' + interfaceName, contract.interface, 'utf-8' )
+                console.log( 'Inteface generated at: ' + parent + "/" + interfaceName )
+            }
             return { deployment_result, contract }
         } catch ( ex ) {
             console.log( ex )
-            if ( ex.response.text ) {
+            if ( ex.response && ex.response.text ) {
                 console.log( JSON.parse( ex.response.text ) )
             }
             throw ex
         }
     }
     const fakeAddress = 'ct_A8WVnCuJ7t1DjAJf4y8hJrAEVpt1T9ypG3nNBdbpKmpthGvUm'
-    const deployments = 
-        [ 
-            /* 00 */ () => deployContract( './contracts/test/BuildAll.aes', [] ),
-            /* 01 */ () => deployContract( './contracts/AedexV2Pair.aes', 
+    const deployments =
+        [
+            /* 00 */ () => deployContract( './contracts/test/BuildAll.aes', []  ),
+            /* 01 */ () => deployContract( './contracts/AedexV2Pair.aes',
                 [ fakeAddress, fakeAddress, fakeAddress, undefined ] ),
-
+            /* 02 */ () => deployContract(
+                './contracts/router/AedexV2Router.aes',
+                [ fakeAddress, fakeAddress, fakeAddress ],
+                'IAedexV2Router.aes',
+            ),
+            /* 03 */ () => deployContract( './contracts/test/WAE.aes', [], 'IWAE.aes' ),
         ]
-    await deployments[0]()
+    await deployments[2]()
 
     //console.log(await contract.methods.getOwner())
 }
