@@ -14,35 +14,18 @@
  *  OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
  *  PERFORMANCE OF THIS SOFTWARE.
  */
-const { assert, expect, use } = require( 'chai' )
-const { BigNumber, BigNumberish, constants, Contract, ContractTransaction, utils, Wallet } =  require( 'ethers' )
-import { Decimal } from 'decimal.js'
-const { jestSnapshotPlugin } = require( "mocha-chai-jest-snapshot" )
-
-const tokenA      = 'ct_A8WVnCuJ7t1DjAJf4y8hJrAEVpt1T9ypG3nNBdbpKmpthGvUm'
-const tokenB      = 'ct_m18VKpSVhsQtjmUC7oZAJPwk36usb39B2viWzBuPjfQEsxHYL'
-const fakeFactory = 'ct_27JMp3z1pyXbfjra2VXiFU9e5jtFTyzus57S6eWbGNh3NSfabo'
-
-use( jestSnapshotPlugin() )
-
+const { expect } = require( 'chai' )
 const { defaultWallets: WALLETS } = require( '../config/wallets.json' )
 
 import {
     getA,
-    getContract,
     pairFixture,
     beforeEachWithSnapshot,
-} from './shared/fixtures.js'
+} from './shared/fixtures'
 
 import {
-    expandTo18Decimals,
-    MaxUint256,
     expectToRevert,
-
-} from './shared/utilities.js'
-
-const TOTAL_SUPPLY = expandTo18Decimals( 10000 )
-const TEST_AMOUNT = expandTo18Decimals( 10 )
+} from './shared/utilities'
 
 const wallet = {
     ...WALLETS[0],
@@ -53,11 +36,11 @@ const other = {
     ...WALLETS[1],
     address: WALLETS[1].publicKey
 }
-var factory, token0, token1, pair
+var factory, token0, token1
 
 describe( 'Pair Factory', () => {
     beforeEachWithSnapshot( 'first compile pool factory', async () => {
-        ( { factory, token0, token1, pair } = await pairFixture() )
+        ( { factory, token0, token1 } = await pairFixture() )
     } )
     it( 'feeTo, feeToSetter, allPairsLength', async () => {
         const exe = factory.exe
@@ -70,7 +53,7 @@ describe( 'Pair Factory', () => {
             () => factory.exe( x => x.create_pair(
                 getA( token0 ),
                 getA( token1 ),
-                getA( factory ),
+                //getA( factory ),
                 undefined
             ) ),
             'AedexV2: PAIR_EXISTS'
@@ -81,7 +64,7 @@ describe( 'Pair Factory', () => {
             () => factory.exe( x => x.create_pair(
                 getA( token1 ),
                 getA( token0 ),
-                getA( factory ),
+                //getA( factory ),
                 undefined
 
             ) ),
